+++ evaluation.py
@@ -1,116 +1,141 @@
-import os
-import json
-import numpy as np
 import pandas as pd
+import seaborn as sns
+import numpy as np
 
-from similarity import compute_similarity
-
-
-# Function to load JSON files from a directory
-def load_json_files(directory):
-    data = {}
-    for file in os.listdir(directory):
-        if file.endswith(".json"):
-            file_path = os.path.join(directory, file)
-            with open(file_path, "r") as f:
-                data[file.replace(".json", "")] = json.load(f)
-    return data
+# ----- Investigation plots -----
 
 
-# Always picks the highest available similarity at each step
-def match_line_items(gt_rows, pred_rows):
-    if not gt_rows or not pred_rows:
-        return {}
+def compute_model_metrics_df(matched_pairs_dict, metric=None, error_type="mae"):
+    from sklearn.metrics import r2_score
 
-    cost_matrix = np.zeros((len(gt_rows), len(pred_rows)))
+    """
+    Computes recall or error metrics (MAE, MAPE, R², Balanced Recall-R²) for different models and examples.
 
-    # Compute similarity matrix
-    for i, gt_row in enumerate(gt_rows):
-        for j, pred_row in enumerate(pred_rows):
-            cost_matrix[i, j] = compute_similarity(gt_row, pred_row)
+    Returns:
+    - A DataFrame containing columns: ["example", "model", "value"].
+    """
+    valid_metrics = ["qty", "rateUsd", "rowTotalCostUsd"]
+    valid_error_types = ["mae", "mape", "r2", "balanced"]
 
-    # Sort pairs by highest similarity score first
-    sorted_pairs = sorted(
-        [
-            (i, j, cost_matrix[i, j])
-            for i in range(len(gt_rows))
-            for j in range(len(pred_rows))
-        ],
-        key=lambda x: -x[2],  # Sort by descending similarity
+    if metric not in valid_metrics:
+        metric = "recall"  # Default to recall if metric is None or invalid
+
+    if error_type not in valid_error_types:
+        error_type = "mae"  # Default to MAE if error_type is invalid
+
+    data = []
+    for model_name, examples in matched_pairs_dict.items():
+        for example_name, example_data in examples.items():
+            df = example_data["matched_pairs_data"]
+            recall = example_data["recall"]
+
+            if metric == "recall":
+                value = recall
+            else:
+                # Compute the chosen error metric
+                gt_col = f"gt_{metric}"
+                pred_col = f"pred_{metric}"
+
+                if gt_col in df.columns and pred_col in df.columns:
+                    gt_values = df[gt_col].values
+                    pred_values = df[pred_col].values
+
+                    if error_type == "mae":
+                        value = np.mean(np.abs(gt_values - pred_values))
+                    elif error_type == "mape":
+                        gt_values = np.where(
+                            gt_values == 0, 1e-6, gt_values
+                        )  # Avoid division by zero
+                        value = (
+                            np.mean(np.abs((gt_values - pred_values) / gt_values)) * 100
+                        )  # In percentage
+                    elif error_type == "r2":
+                        value = r2_score(gt_values, pred_values)
+                    elif error_type == "balanced":
+                        r2_value = r2_score(gt_values, pred_values)
+                        r2_normalized = (
+                            r2_value + 1
+                        ) / 2  # Normalize R² to (0,1) range
+                        value = (
+                            np.sqrt(recall * r2_normalized)
+                            if recall is not None and r2_value is not None
+                            else None
+                        )
+                else:
+                    value = None  # Handle missing columns
+
+            data.append({"example": example_name, "model": model_name, "value": value})
+
+    return pd.DataFrame(data)
+
+
+def aggregate_by_model(df):
+    """
+    Aggregates metric values by model, computing the mean across all examples.
+
+    Parameters:
+    - df: DataFrame containing ["example", "model", "value"].
+
+    Returns:
+    - plot_data: Aggregated DataFrame with mean values per model.
+    - model_order: Sorted list of model names.
+    """
+    plot_data = df.groupby("model", as_index=False)[
+        "value"
+    ].mean()  # Aggregate over examples
+    model_order = sorted(
+        plot_data["model"].unique(), key=lambda x: int(x)
+    )  # Order models numerically
+
+    return plot_data, model_order
+
+
+def plot_model_metrics(df, x="example"):
+    import matplotlib.pyplot as plt
+
+    """
+    Plots a DataFrame containing model evaluation metrics.
+
+    Parameters:
+    - df: DataFrame containing ["example", "model", "value"].
+    - x: "example" to plot per example, or "model" to aggregate over all examples.
+    """
+    valid_x_types = ["example", "model"]
+    if x not in valid_x_types:
+        x = "example"  # Default to example-based plotting
+
+    # Sort x-axis and hue before plotting
+    if x == "example":
+        x_label = "Example"
+        example_order = sorted(df["example"].unique())  # Order examples alphabetically
+        model_order = sorted(
+            df["model"].unique(), key=lambda x: int(x)
+        )  # Order models numerically
+        plot_data = df
+    else:  # Aggregate by model
+        x_label = "Model"
+        plot_data, model_order = aggregate_by_model(df)
+
+    # Create bar plot
+    plt.figure(figsize=(12, 6))
+    sns.barplot(
+        data=plot_data,
+        x=x,
+        y="value",
+        hue=(
+            "model" if x == "example" else None
+        ),  # Use hue for models only if plotting by example
+        order=example_order if x == "example" else model_order,
     )
 
-    matched_pairs = {}
-    used_gt = set()
-    used_pred = set()
+    # Formatting
+    plt.xlabel(x_label)
+    plt.ylabel("Metric Value")
+    plt.title(f"Model Metrics by {x_label}")
+    plt.xticks(rotation=45)
 
-    # Greedily match the highest similarity pairs first
-    for gt_idx, pred_idx, _ in sorted_pairs:
-        if gt_idx not in used_gt and pred_idx not in used_pred:
-            matched_pairs[gt_idx] = pred_idx
-            used_gt.add(gt_idx)
-            used_pred.add(pred_idx)
+    if x == "example":
+        plt.legend(title="Model")
 
-    return matched_pairs
-
-
-# Step 2: Compare Numerical Metrics (MAPE Calculation)
-def calculate_mape(gt_rows, pred_rows, matched_pairs):
-    qty_mape, rate_mape, cost_mape = [], [], []
-
-    for gt_idx, pred_idx in matched_pairs.items():
-        gt_row = gt_rows[gt_idx]
-        pred_row = pred_rows[pred_idx]
-
-        if gt_row["qty"] > 0:
-            qty_mape.append(abs(gt_row["qty"] - pred_row["qty"]) / gt_row["qty"])
-        if gt_row["rateUsd"] > 0:
-            rate_mape.append(
-                abs(gt_row["rateUsd"] - pred_row["rateUsd"]) / gt_row["rateUsd"]
-            )
-        if gt_row["rowTotalCostUsd"] > 0:
-            cost_mape.append(
-                abs(gt_row["rowTotalCostUsd"] - pred_row["rowTotalCostUsd"])
-                / gt_row["rowTotalCostUsd"]
-            )
-
-    results = {
-        "Qty MAPE": np.mean(qty_mape) if qty_mape else None,
-        "Rate MAPE": np.mean(rate_mape) if rate_mape else None,
-        "Cost MAPE": np.mean(cost_mape) if cost_mape else None,
-    }
-    return results
-
-
-# Evaluate a Single Model
-def evaluate_model(model_data, ground_truth_data):
-    qty_mape, rate_mape, cost_mape = [], [], []
-
-    for prediction in model_data["estimate_preds"]:
-        file_name = prediction["valid_file_name"]
-        if file_name not in ground_truth_data:
-            continue
-
-        gt_rows = ground_truth_data[file_name]["rows"]
-        pred_rows = prediction["rows"]
-
-        matched_pairs = match_line_items(gt_rows, pred_rows)
-        mape_results = calculate_mape(gt_rows, pred_rows, matched_pairs)
-
-        qty_mape.append(mape_results["Qty MAPE"])
-        rate_mape.append(mape_results["Rate MAPE"])
-        cost_mape.append(mape_results["Cost MAPE"])
-
-    return {
-        "Qty MAPE": np.mean([x for x in qty_mape if x is not None]),
-        "Rate MAPE": np.mean([x for x in rate_mape if x is not None]),
-        "Cost MAPE": np.mean([x for x in cost_mape if x is not None]),
-    }
-
-
-# Evaluate All Models Using the Modular Approach
-def evaluate_all_models(model_output_data, ground_truth_data):
-    model_results = {
-        model_name: evaluate_model(model_data, ground_truth_data)
-        for model_name, model_data in model_output_data.items()
-    }
-    return pd.DataFrame.from_dict(model_results, orient="index")
+    # Show plot
+    plt.show()
